convert spdx structs to versioned pkgs
