+++ spdx/common/identifier.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0 OR GPL-2.0-or-later
 
-package spdx
+package common
 
 import (
 	"encoding/json"
